+++ examples/queries/src/query.rs
@@ -24,9 +24,7 @@ struct Args {
 pub struct AgentProps {}
 
 fn query_callback(_ctx: &Arc<Context>, _props: &Arc<RwLock<AgentProps>>, answer: &[u8]) {
-	let config = bincode::config::standard();
-	let (message, _len): (String, usize) =
-		bincode::decode_from_slice(answer, config).expect("should not happen");
+	let message: String =	bitcode::decode(answer).expect("should not happen");
 	println!("Received '{}'", &message);
 }